Moved exceptions to separate package
