+++ Server/src/server/command/LoginCommand.java
@@ -1,7 +1,7 @@
 package server.command;
 
 import com.google.gson.annotations.SerializedName;
-import server.BadCredentialsException;
+import server.exception.BadCredentialsException;
 import server.ServerFacade;
 import server.responses.Response;
 import server.responses.ResponseWrapper;
