+++ Server/src/server/exception/InternalServerException.java
@@ -1,4 +1,4 @@
-package server;
+package server.exception;
 
 /**
  *
